requirements.txt
bootstrap4==0.1.0
certifi==2019.6.16
chardet==3.0.4
config==0.3.9
dj-database-url==0.5.0
Django==2.2.5
django-celery-beat==1.5.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-filter==2.2.0
django-heroku==0.3.1
django-paypal==1.0.0
django-richtextfield==1.4.0
django-timezone-field==3.0
gunicorn==19.9.0
idna==2.8
Pillow==6.1.0
psycopg2==2.8.3
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
stripe==2.33.2
urllib3==1.25.3
whitenoise==4.1.3
